dom/media/autoplay/test/mochitest/file_autoplay_policy_key_blacklist.html

<!DOCTYPE HTML>
<html>

<head>
  <title>Autoplay policy window</title>
  <style>
    video {
      width: 50%;
      height: 50%;
    }
    :focus {
      background-color: blue;
    }
  </style>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="manifest.js"></script>
  <script type="text/javascript" src="AutoplayTestUtils.js"></script>
</head>

<body>
  <div id="x">This is a div with id=x.</div>
  <pre id="test">
      <input type="text" id="text-input"/>
      <script>

        window.ok = window.opener.ok;
        window.is = window.opener.is;
        window.info = window.opener.info;

        // Keys that are expected to be not considered interaction with the page, and
        // so not gesture activate the document.
        let blacklistKeyPresses = [
          "CapsLock",
          "NumLock",
          "ScrollLock",
          "FnLock",
          "Meta",
          "ArrowUp",
          "ArrowDown",
          "ArrowLeft",
          "ArrowRight",
          "PageUp",
          "PageDown",
          "Home",
          "End",
          "Fn",
          "Alt",
          "AltGraph",
          "Control",
          "Shift",
          "Escape",
        ];

        // XXX not sure why Android behave different on handling Backspace key.
        if (!navigator.appVersion.includes("Android")) {
          blacklistKeyPresses.push("Backspace");
        }

        let modifiedKeys = [
          { key: "a", modifiers: { accelKey: true } },
          { key: "z", modifiers: { accelKey: true } },
        ];

        // Browser shortcut on non-Linux platform.
        if (navigator.platform.indexOf("Linux") !== 0) {
          modifiedKeys.push(
            { key: "KEY_ArrowRight", modifiers: { metaKey: true } },
            { key: "KEY_ArrowRight", modifiers: { altKey: true } });
        }

        let script = SpecialPowers.loadChromeScript(function() {
          /* eslint-env mozilla/chrome-script */
          var EventUtils = {};
          EventUtils.window = {};
          EventUtils._EU_Ci = Ci;
          EventUtils._EU_Cc = Cc;
          Services.scriptloader
            .loadSubScript("chrome://mochikit/content/tests/SimpleTest/EventUtils.js",
                           EventUtils);

          addMessageListener("synthesizeKey", function(data) {
            let browserWin = Services.wm.getMostRecentWindow("navigator:browser");
            if (!browserWin) {
              browserWin = Services.wm.getMostRecentWindow("navigator:geckoview");
            }
            EventUtils.synthesizeKey(data.key, data.event, browserWin);
          });
        });

        async function synthesizeKeyAndWait(key, event = {}) {
          let promise = new Promise(aResolve => {
            document.addEventListener(
              "keydown",
              function (e) {
                e.preventDefault();
                info("Received keydown event: " + e.key);
                aResolve();
              },
              { once: true }
            );
          });
          script.sendAsyncMessage("synthesizeKey", { key, event });
          await promise;
        }

        async function sendInput(element, name, input) {
          synthesizeMouseAtCenter(input, {});
          let played = await element.play().then(() => true, () => false);
          ok(!played, "Clicking " + name + " should not activate document and should not unblock play");

          synthesizeCompositionChange({
            composition: {
              string: "\u30E9\u30FC\u30E1\u30F3",
              clauses: [
                { length: 4, attr: COMPOSITION_ATTR_RAW_CLAUSE }
              ]
            },
            caret: { start: 4, length: 0 }
          });
          synthesizeComposition({ type: "compositioncommitasis" });
          played = await element.play().then(() => true, () => false);
          ok(!played, "Entering text to " + name + " via IME should not activate document and should not unblock play");

          input.focus();
          sendString("ascii text");
          played = await element.play().then(() => true, () => false);
          ok(!played, "Entering ASCII text into " + name + " should not activate document and should not unblock play");

          input.blur();
        }

        async function testAutoplayKeyBlacklist() {
          let element = document.createElement("video");
          element.preload = "auto";
          element.src = "short.mp4";
          document.body.appendChild(element);

          await once(element, "loadedmetadata");

          let played = await element.play().then(() => true, () => false);
          is(played, false, "Document should start out not activated, with playback blocked.");

          // Try pressing all the keys in the blacklist, then playing.
          // Document should not be activated, so play should fail.

          for (let key of blacklistKeyPresses) {
            document.body.focus();
            await synthesizeKeyAndWait("KEY_" + key);
            played = await element.play().then(() => true, () => false);
            is(played, false, "Key " + key + " should not activate document and should not unblock play");
          }

          // Try pressing some keys with modifiers.
          let keyNames = (m) => Object.keys(m).join("+");
          for (let x of modifiedKeys) {
            document.body.focus();
            await synthesizeKeyAndWait(x.key, x.modifiers);
            played = await element.play().then(() => true, () => false);
            is(played, false, "Key (" + x.key + "+" + keyNames(x.modifiers) + ") should not activate document and should not unblock play");
          }

          // Try pressing a key not in the blacklist, then playing.
          // Document should be activated, and media should play.
          await synthesizeKeyAndWait(" ");
          played = await element.play().then(() => true, () => false);
          is(played, true, "Space key should activate document and should unblock play");

          removeNodeAndSource(element);
        }

        nextWindowMessage().then(
          async (event) => {
            try {
              await testAutoplayKeyBlacklist(event.data, event.source);
            } catch (e) {
              ok(false, "Caught exception " + e + " " + e.message + " " + e.stackTrace);
            }
            event.source.postMessage("done", "*");
          });

      </script>
    </pre>
</body>

</html>
